server/pyproject.toml
[tool.poetry]
name = "grpc-server"
version = "1"
description = "grpc python server"
authors = ["Your Name <you@example.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.62.2"
grpcio-tools = "^1.62.2"
grpc-interceptor = "^0.15.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
